+++ 4-storage/redshift-storage/sql/com.google.analytics.enhanced-ecommerce/impression_field_object_1.sql
@@ -22,7 +22,7 @@ CREATE TABLE IF NOT EXISTS atomic.com_google_analytics_enhanced_ecommerce_impres
     "id"             VARCHAR(500),
     "list"           VARCHAR(500),
     "name"           VARCHAR(500),
-    "position"       BIGINT,
+    "position"       INT,
     "price"          dec(36,2),
     "variant"        VARCHAR(500),
     FOREIGN KEY (root_id) REFERENCES atomic.events(event_id)
